Moved tests outside src
