+++ __tests__/GetRouteHandler.spec.ts
@@ -1,5 +1,5 @@
-import { UserController } from './__mocks__/UserController.mock'
-import { BindController } from './BindController'
+import { UserController } from '../__mocks__/UserController.mock'
+import { BindController } from '../src/BindController'
 
 const userController = new UserController()
