--- src/Router.spec.ts
@@ -1,3 +1,0 @@
-describe('Router', () => {
-  it('should work', () => expect(1).toBe(1))
-})
