+++ tsconfig.json
@@ -9,5 +9,5 @@
     "esModuleInterop": true
   },
   "include": ["src/**/*"],
-  "exclude": ["node_modules", "**/*-spec.ts", "src/__tests__/*", "src/mocks/*"]
+  "exclude": ["node_modules", "**/*-spec.ts", "src/__tests__/*", "src/__mocks__/*"]
 }
